Move the AnalysisManager to its own module.
It is, so far, unchanged except for things required to make it work from
a file separate from scheduler.py.
